server: move diagnostics to its own directory in service
